+++ src/Funds/utils/index.js
@@ -1,0 +1,1 @@
+export { getPlannedFiscalYears } from './getPlannedFiscalYears';